Convert into a lib project with an external main
This will allow using the same rebase logic for other binaries.
